+++ tests/TestCase.php
@@ -2,9 +2,7 @@
 
 namespace Spatie\Honeypot\Tests;
 
-use Carbon\CarbonInterface;
 use Illuminate\Foundation\Testing\Concerns\InteractsWithContainer;
-use Illuminate\Support\Facades\Date;
 use Illuminate\Support\Facades\View;
 use Spatie\Honeypot\HoneypotServiceProvider;
 use Spatie\Honeypot\Tests\TestClasses\FakeEncrypter;
@@ -21,26 +19,11 @@ abstract class TestCase extends \Orchestra\Testbench\TestCase
 
         View::addLocation(__DIR__.'/views');
 
-        if ($this->testNow) {
-            $this->setNow(2019, 1, 1);
-        }
-
         $this->swap('encrypter', new FakeEncrypter());
     }
 
     protected function getPackageProviders($app)
     {
         return [HoneypotServiceProvider::class];
-    }
-
-    protected function setNow($year, int $month = 1, int $day = 1)
-    {
-        $newNow = $year instanceof CarbonInterface
-            ? $year->copy()
-            : Date::createFromDate($year, $month, $day);
-
-        $newNow = $newNow->startOfDay();
-
-        Date::setTestNow($newNow);
     }
 }