+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `laravel-honeypot` will be documented in this file
 
+## 3.X.X - unreleased
+
+- drop support for anything below PHP 7.4, Laravel 8
+- add `x-honeypot` Blade component
+- move setup to dedicated class, so it can be used in Inertia
+
 ## 2.3.0 - 2020-09-08
 
 - add support for Laravel 8